+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/Dynatrace/openkit-native/compare/v3.0.0...HEAD)
 
+### Changed
+
+- Non-finite numeric values are serialized as JSON null in reported events, and a special field is added for supportability.
+
 ## 3.0.0 [Release date: 2022-12-07]
 [GitHub Releases](https://github.com/Dynatrace/openkit-native/releases/tag/v3.0.0)